requirements.txt
marshmallow
mock
pyjwt[crypto]
pytest
python-dateutil
python-i18n[YAML]
pytz
requests
requests-mock
twine
